Refactoring mc
- Refactoring mc command
-  You can now check other mc servers
- Started refactoring coinflip
