+++ models/User.js
@@ -1,0 +1,19 @@
+module.exports = (sequelize, DataTypes) => {
+  return sequelize.define(
+    "users",
+    {
+      user_id: {
+        type: DataTypes.STRING,
+        primaryKey: true,
+      },
+      balance: {
+        type: DataTypes.INTEGER,
+        defaultValue: 100,
+        allowNull: false,
+      },
+    },
+    {
+      timestamps: false,
+    }
+  );
+};
